Module-11-LangChain-Framework/03_ChatPromptTemplate/requirements.txt
gradio
langchain-core
langchain-openai
python-dotenv
langchain-huggingface
numpy
